+++ src/hklpy2/tests/test_diffract.py
@@ -22,10 +22,10 @@ from ..blocks.sample import Sample
 from ..diffract import DiffractometerBase
 from ..diffract import creator
 from ..diffract import diffractometer_class_factory
-from ..misc import ConfigurationError
-from ..misc import DiffractometerError
-from ..misc import NoForwardSolutions
-from ..misc import VirtualPositionerBase
+from ..exceptions import ConfigurationError
+from ..exceptions import DiffractometerError
+from ..exceptions import NoForwardSolutions
+from ..devices import VirtualPositionerBase
 from ..ops import DEFAULT_SAMPLE_NAME
 from ..ops import Core
 from ..ops import CoreError